tests/wpt/tests/css/css-grid/grid-lanes/subgrid/grid-subgridded-to-grid-lanes/track-sizing/row-subgrid-auto-repeat-auto-001.html

<!DOCTYPE html>
<html>
<title>CSS Grid Lanes Test: intrinsic auto-repeat rows sized by a subgrid item</title>
<link rel="help" href="https://drafts.csswg.org/css-grid-3/#subgrids">
<link rel="author" title="Alison Maher" href="mailto:almaher@microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.grid-lanes {
    display: grid-lanes;
    flow-tolerance: 0;
    grid-template-rows: repeat(auto-fill, auto);
    height: 300px;
    width: 100px;
}

.grid-lanes > div {
    width: 100px;
    height: 100px;
}

.subgrid {
    display: grid;
    grid-template-rows: subgrid;
}

.subgrid > div {
    width: 100px;
    height: 100px;
}
</style>
<div class="grid-lanes">
  <div></div>
  <div></div>
  <div class="subgrid" style="grid-row: auto / span 1;">
    <div></div>
  </div>
</div>
<script>
  test(function() {
    const container = document.querySelector('.grid-lanes');
    const computedStyle = window.getComputedStyle(container);
    assert_equals(computedStyle.getPropertyValue('grid-template-rows'),
      "100px 100px 100px");
  });
</script>
</html>
